feat: AuthenticationHandler + remove annotation injection in Main
